BUT SEMESTRE 2 _ SAE/SAE2.03 - Mettre en place une solution informatique pour l_entreprise/requirements.txt
flask
mysql-connector-python
bcrypt


